# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses/{courseId}/assessments:
    get:
      operationId: listCourseAssessments
      tags: [assessments]
      summary: List practice assessments
      parameters:
        - in: path
          name: courseId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Assessments
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AssessmentSummary" }
    post:
      operationId: createCourseAssessment
      tags: [assessments]
      summary: Generate a source-grounded practice assessment
      parameters:
        - in: path
          name: courseId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AssessmentInput" }
      responses:
        "201":
          description: Assessment created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AssessmentDetail" }
        "404": { description: Course or topic not found }
        "409": { description: Course has no ready study units }
  /assessments/{assessmentId}:
    get:
      operationId: getAssessment
      tags: [assessments]
      summary: Get an assessment without revealing its answer key
      parameters:
        - in: path
          name: assessmentId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Assessment
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AssessmentDetail" }
        "404": { description: Assessment not found }
  /assessments/{assessmentId}/attempts:
    get:
      operationId: listAssessmentAttempts
      tags: [assessments]
      summary: List the signed-in learner's attempts
      parameters:
        - in: path
          name: assessmentId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Attempts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AssessmentAttempt" }
        "404": { description: Assessment not found }
    post:
      operationId: submitAssessmentAttempt
      tags: [assessments]
      summary: Grade a completed practice assessment
      parameters:
        - in: path
          name: assessmentId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AssessmentAnswersInput" }
      responses:
        "201":
          description: Graded attempt with explanations and citations
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AssessmentAttempt" }
        "404": { description: Assessment not found }
...
components:
  schemas:
...
    SourceCitation:
      type: object
      properties:
        unitId: { type: string }
        sourceId: { type: string }
        sourceTitle: { type: string }
        sourceType: { type: string, enum: [pdf, pptx, video] }
        locator:
          type: object
          properties:
            type: { type: string, enum: [page, slide, video] }
            page: { type: integer, minimum: 1 }
            slide: { type: integer, minimum: 1 }
            start_sec: { type: number, minimum: 0 }
            end_sec: { type: number, minimum: 0 }
            imagePath: { type: ["string", "null"] }
          required: [type]
        excerpt: { type: string }
      required: [unitId, sourceId, sourceTitle, sourceType, locator, excerpt]
...
    AssessmentInput:
      type: object
      properties:
        questionCount: { type: integer, minimum: 2, maximum: 15, default: 5 }
        difficulty: { type: string, enum: [easy, medium, hard], default: medium }
        topicId: { type: ["string", "null"] }
      required: [questionCount, difficulty]
    AssessmentQuestion:
      type: object
      properties:
        id: { type: string }
        prompt: { type: string }
        options:
          type: array
          minItems: 2
          items: { type: string }
        citations:
          type: array
          items: { $ref: "#/components/schemas/SourceCitation" }
      required: [id, prompt, options, citations]
    AssessmentSummary:
      type: object
      properties:
        id: { type: string }
        courseId: { type: string }
        title: { type: string }
        difficulty: { type: string, enum: [easy, medium, hard] }
        questionCount: { type: integer }
        attemptCount: { type: integer }
        bestScore: { type: ["number", "null"] }
        createdAt: { type: string, format: date-time }
      required:
        [id, courseId, title, difficulty, questionCount, attemptCount, bestScore, createdAt]
    AssessmentDetail:
      allOf:
        - $ref: "#/components/schemas/AssessmentSummary"
        - type: object
          properties:
            questions:
              type: array
              items: { $ref: "#/components/schemas/AssessmentQuestion" }
          required: [questions]
    AssessmentAnswersInput:
      type: object
      properties:
        answers:
          type: array
          minItems: 1
          items:
            type: object
            properties:
              questionId: { type: string }
              selectedOption: { type: ["integer", "null"], minimum: 0 }
            required: [questionId, selectedOption]
      required: [answers]
    AssessmentQuestionFeedback:
      type: object
      properties:
        questionId: { type: string }
        selectedOption: { type: ["integer", "null"], minimum: 0 }
        correctOption: { type: integer, minimum: 0 }
        isCorrect: { type: boolean }
        explanation: { type: string }
        citations:
          type: array
          items: { $ref: "#/components/schemas/SourceCitation" }
      required: [questionId, selectedOption, correctOption, isCorrect, explanation, citations]
    AssessmentAttempt:
      type: object
      properties:
        id: { type: string }
        assessmentId: { type: string }
        score: { type: integer, minimum: 0 }
        total: { type: integer, minimum: 1 }
        percentage: { type: number, minimum: 0, maximum: 100 }
        feedback:
          type: array
          items: { $ref: "#/components/schemas/AssessmentQuestionFeedback" }
        createdAt: { type: string, format: date-time }
      required: [id, assessmentId, score, total, percentage, feedback, createdAt]
...